refactor content script to not use Class structure; split out video player controls functionality into separate file
